Adds best_defense method and passing test. Moves methods to the correct module.
