+++ lib/league.rb
@@ -65,6 +65,123 @@ module League
     results.min_by { |team| team[1][:average] }.first
   end
 
+  # Description: Name of the team with the highest average number of goals allowed per game across all seasons
+    def worst_defense
+      team_ids_hash = Hash.new
+      @teams.each do |team|
+        team_ids_hash[team[:team_id]] = {goals_against: 0, games: 0, average: 0}
+      end
+
+      @combine_data.each do |game|
+        home_team_id = team_ids_hash[game[:home_team_id]]
+        home_team_id[:goals_against] += game[:away_goals].to_i
+        home_team_id[:games] += 1
+        home_team_id[:average] = (home_team_id[:goals_against].to_f / home_team_id[:games]).round(3)
+
+        away_team_id = team_ids_hash[game[:away_team_id]]
+        away_team_id[:goals_against] += game[:home_goals].to_i
+        away_team_id[:games] += 1
+        away_team_id[:average] = (away_team_id[:goals_against].to_f / away_team_id[:games]).round(3)
+      end
+      worst_defense_id = team_ids_hash.max_by { |team| team[1][:average] }[0]
+      worst_defense = @teams.find do |team|
+        team[:team_id] == worst_defense_id
+      end
+      worst_defense[:team_name]
+    end
+
+  # Description: Name of the team with the lowest average number of goals allowed per game across all seasons
+    def best_defense
+      team_ids_hash = Hash.new
+      @teams.each do |team|
+        team_ids_hash[team[:team_id]] = {goals_against: 0, games: 0, average: 0}
+      end
+
+      @combine_data.each do |game|
+        home_team_id = team_ids_hash[game[:home_team_id]]
+        home_team_id[:goals_against] += game[:away_goals].to_i
+        home_team_id[:games] += 1
+        home_team_id[:average] = (home_team_id[:goals_against].to_f / home_team_id[:games]).round(3)
+
+        away_team_id = team_ids_hash[game[:away_team_id]]
+        away_team_id[:goals_against] += game[:home_goals].to_i
+        away_team_id[:games] += 1
+        away_team_id[:average] = (away_team_id[:goals_against].to_f / away_team_id[:games]).round(3)
+      end
+
+      teams_with_data = team_ids_hash.delete_if { |k,v| v[:average] <= 0 }
+      best_defense_id = teams_with_data.min_by { |team| team[1][:average] }[0]
+      best_defense = @teams.find do |team|
+        team[:team_id] == best_defense_id
+      end
+      best_defense[:team_name]
+    end
+
+    # Description: Name of the team with the highest win percentage across all seasons
+    # Return Value: String
+    def winningest_team
+      team_ids_hash = Hash.new
+      @teams.each do |team|
+        team_ids_hash[team[:team_id]] = {games_won: 0, games: 0, average: 0}
+      end
+
+      @game_teams.each do |game|
+        team_id = team_ids_hash[game[:team_id]]
+        team_id[:games_won] += 1 if game[:won] == "TRUE"
+        team_id[:games] += 1
+        team_id[:average] = (team_id[:games_won].to_f / team_id[:games]).round(3)
+      end
+
+      winningest_id = team_ids_hash.max_by { |team| team[1][:average] }[0]
+      winningest_team = @teams.find do |team|
+        team[:team_id] == winningest_id
+      end
+
+      winningest_team[:team_name]
+    end
+
+    # Iteration 3
+    # Description: Name of the team with the highest average score per game across all seasons when they are home.
+    # Return Value: String
+    def highest_scoring_home_team
+      team_ids_hash = Hash.new
+      @teams.each do |team|
+        team_ids_hash[team[:team_id]] = {goals: 0, games: 0, average: 0}
+      end
+
+      @combine_data.each do |game|
+        team_id = team_ids_hash[game[:home_team_id]]
+        team_id[:goals] += game[:home_goals].to_i
+        team_id[:games] += 1
+        team_id[:average] = (team_id[:goals].to_f / team_id[:games]).round(2)
+      end
+      high_scoring_home_id = team_ids_hash.max_by { |team| team[1][:average] }[0]
+
+      high_scoring_home_team = @teams.find do |team|
+        team[:team_id] == high_scoring_home_id
+      end
+      high_scoring_home_team[:team_name]
+    end
+
+    # Iteration 3
+    # Description: Name of the team with the highest average number
+    # of goals scored per game across all seasons.
+    # Return Value: String
+    def best_offense
+      team_names_hash = Hash.new
+      @teams.each do |team|
+        team_names_hash[team[:team_name]] = {goals: 0, games: 0, average: 0}
+      end
+
+      @combine_data.each do |game|
+        team_name = team_names_hash[game[:team_name]]
+        team_name[:goals] += game[:goals].to_i
+        team_name[:games] += 1
+        team_name[:average] = (team_name[:goals].to_f / team_name[:games]).round(2)
+      end
+      team_names_hash.max_by { |team| team[1][:average] }[0]
+    end
+
   # def goals_allowed
   #   allowed = Hash.new { |hash, key| hash[key] = [] }
   #   @combine_data.each do |game|
