+++ lib/team_statistics.rb
@@ -1,6 +1,6 @@
 module TeamStatistics
   def team_info(team_id)
-    results = @teams.find { |team| team[:team_id] == team_id }
+    results = @teams.find { |team| team[:team_id] == team_id.to_i }
     {
       'team_id' => results[:team_id].to_s,
       'franchise_id' => results[:franchise_id].to_s,
