+++ test/test_helper.rb
@@ -1,5 +1,5 @@
-require 'simplecov'
-SimpleCov.start
+# require 'simplecov'
+# SimpleCov.start
 require 'minitest/autorun'
 require 'minitest/emoji'
 require './lib/stat_tracker'
